requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0.7
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
mod-wsgi==4.6.4+ap24vc14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3
six==1.14.0
tablib==1.1.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
